+++ examples/cpi/cpi-metaplex-7/trident-tests/fuzz_0/test_fuzz.rs
@@ -27,7 +27,7 @@ fn main() {
         processor!(entry),
     );
 
-    let config = Config::new();
+    let config = TridentConfig::new();
     let mut client = TridentSVM::new_client(&[program], &config);
-    fuzz_trident ! (fuzz_ix : FuzzInstruction , | fuzz_data : InstructionsSequence , client : TridentSVM , config : Config |);
+    fuzz_trident ! (fuzz_ix : FuzzInstruction , | fuzz_data : InstructionsSequence , client : TridentSVM , config : TridentConfig |);
 }